refactor: modify util helpers location
